fix debug output by casting currentRule from rule_t to int

+++ lpd8806_life.pde
@@ -132,7 +132,7 @@ void loop()
     resetBufferAges();
     
     // debugging for funsies
-    Serial.println(currentRule);
+    Serial.println( (int) currentRule );
   }
 
   caStepNumber++;
